TheAI/src/compliance_reporting.py
```python
import pandas as pd
from datetime import datetime

class ComplianceReport:
    def __init__(self):
        self.compliance_report = pd.DataFrame(columns=['Date', 'Threats Detected', 'Firewall Configurations', 'Virus Scans', 'Intrusions Detected', 'Penetration Tests', 'Real Time Threats Detected', 'Employee Trainings', 'Self Destruct Activations', 'IoT Connections'])

    def update_report(self, threat_prediction_model, firewall_configuration, virus_database, intrusion_detection_system, pen_test_results, real_time_threat_detection_system, employee_training_material, self_destruct_sequence, iot_compatibility_status):
        new_entry = {
            'Date': datetime.now(),
            'Threats Detected': len(threat_prediction_model.predictions),
            'Firewall Configurations': len(firewall_configuration.configurations),
            'Virus Scans': len(virus_database.scans),
            'Intrusions Detected': len(intrusion_detection_system.detections),
            'Penetration Tests': len(pen_test_results.tests),
            'Real Time Threats Detected': len(real_time_threat_detection_system.detections),
            'Employee Trainings': len(employee_training_material.trainings),
            'Self Destruct Activations': len(self_destruct_sequence.activations),
            'IoT Connections': len(iot_compatibility_status.connections)
        }
        self.compliance_report = self.compliance_report.append(new_entry, ignore_index=True)

    def generate_report(self):
        return self.compliance_report

compliance_report = ComplianceReport()
```
