TheAI/tests/test_compliance_reporting.py
```python
import unittest
from TheAI.src.compliance_reporting import generate_report

class TestComplianceReporting(unittest.TestCase):

    def setUp(self):
        self.report = generate_report()

    def test_report_generation(self):
        self.assertIsNotNone(self.report, "Report generation failed.")

    def test_report_content(self):
        self.assertIn('compliance_status', self.report, "Compliance status not found in report.")
        self.assertIn('non_compliant_areas', self.report, "Non-compliant areas not found in report.")
        self.assertIn('recommendations', self.report, "Recommendations not found in report.")

    def test_compliance_status(self):
        self.assertIsInstance(self.report['compliance_status'], bool, "Compliance status should be a boolean.")

    def test_non_compliant_areas(self):
        self.assertIsInstance(self.report['non_compliant_areas'], list, "Non-compliant areas should be a list.")

    def test_recommendations(self):
        self.assertIsInstance(self.report['recommendations'], list, "Recommendations should be a list.")

if __name__ == '__main__':
    unittest.main()
```